refactor(handlers): extract lint.ts handler from intent.ts (Stage B)
handleSAPLint (offline abaplint + ADT formatter). 3672 tests, A1 zero-diff.
